backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.20.0",
    "bcrypt>=4.0.0",
    "chromadb>=1.4.1",
    "filelock>=3.20.3",
    "fastapi>=0.128.0",
    "httpx>=0.27.0",
    "llama-index>=0.14.13",
    "openai>=1.0.0",
    "pymupdf>=1.26.4",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "simpleeval>=1.0.3",
    "sqlalchemy[asyncio]>=2.0.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.40.0",
    "alembic>=1.18.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["data*", "tests*"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]
addopts = "-q"
asyncio_mode = "auto"
